requirements.txt
fastapi
uvicorn
requests
pytest
allure-pytest
sqlalchemy
psycopg2-binary
